Tests/LibWeb/Text/input/DOM/Window-methods-that-must-do-nothing.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script>
    test(() => {
        println(`window.captureEvents() returns undefined: ${window.captureEvents() === undefined}`);
        println(`window.releaseEvents() returns undefined: ${window.releaseEvents() === undefined}`);
    });
</script>
