Tests/LibWeb/Text/expected/DOM/Window-methods-that-must-do-nothing.txt

window.captureEvents() returns undefined: true
window.releaseEvents() returns undefined: true
